Factored out classes related to creation of the training material into training_storage module
